requirements.txt
# Core dependencies
elasticsearch>=8.0.0,<9.0.0
elasticsearch[async]>=8.0.0,<9.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
faker>=18.0.0
tqdm>=4.65.0

# Web application
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.24.0

# Async support
aiofiles>=23.0.0

# CLI utilities
click>=8.1.0
rich>=13.0.0

# Data processing
orjson>=3.9.0
